hot reload of docker container script added

+++ hotReloadDocker.sh
@@ -1,15 +1,24 @@
 echo "DOCKER HOT RELOAD PROCESS.. .."
-container_id=$(docker ps --filter ancestor=x:v1 --format "{{.ID}}")
-if [ -n "${container_id}" ]; then
-    echo "Container exists: $container_id"
-    docker stop "$container_id"
-    echo "Stopped container: $container_id"
+image_id=$(docker images --filter "reference=x:v1"  --format "{{.ID}}")
+echo $image_id
+if [ -n "${image_id}" ]; then
+    container_id=$(docker ps --filter ancestor=x:v1 --format "{{.ID}}")
+    if [ -n "${container_id}" ]; then
+        docker stop "$container_id"
+        echo "$container_id stopped..."
+        docker rm "$container_id"
+        echo "$container_id removed..."
+    else
+        docker rmi x:v1 -f
+        echo "removed image"
+        
+    fi
+else 
+    echo "image id not found"
 fi
-docker rmi x:v1 -f
-echo "Removed image"
+
 docker build --network=host -t x:v1 .
 echo "Built image"
 docker run -p 5000:5000 x:v1
 echo "Running container"
 
-#docker images --filter "reference=x:v1" --format "{{.Repository}}:{{.Tag}} => {{.ID}}"
